switch byte int

+++ src/main/java/isep/web/sakila/jpa/entities/Category.java
@@ -18,7 +18,7 @@ public class Category implements Serializable {
 
 	@Id
 	@Column(name="category_id", unique=true, nullable=false)
-	private byte categoryId;
+	private int categoryId;
 
 	@Column(name="last_update", nullable=false)
 	private Timestamp lastUpdate;
